requirements.txt
pytest-playwright==0.5.0
pytest-xdist==3.6.1
allure-pytest==2.13.5
python-dotenv==1.0.1
mypy==1.10.1